+++ src/client/actions/index.js
@@ -1,4 +1,5 @@
 export const FETCH_LEARNERS = 'fetch_learners';
+export const SET_SKILLS = 'set_skills';
 export const DONE_LOADING = 'done_loading';
 export const SEARCH_BY_SKILL = 'search_by_skill';
 export const SEARCH_BY_NAME = 'search_by_name';
@@ -8,6 +9,13 @@ export function fetchLearners(allLearners) {
     type: FETCH_LEARNERS,
     payload: allLearners,
     loading: true,
+  };
+}
+
+export function setSkills(allSkills) {
+  return {
+    type: SET_SKILLS,
+    skills: allSkills,
   };
 }
