+++ src/client/reducers/reducer-learner.jsx
@@ -1,6 +1,7 @@
 import {
   FETCH_LEARNERS,
   DONE_LOADING,
+  SET_SKILLS,
   SEARCH_BY_SKILL,
   SEARCH_BY_NAME
 } from '../actions/';
@@ -12,10 +13,17 @@ export default function(state = { loading: true }, action) {
         learners: action.payload,
         loading: action.loading,
       };
+    case SET_SKILLS:
+      return {
+        learners: state.learners,
+        loading: state.loading,
+        skills: action.skills,
+      };
     case DONE_LOADING:
       return {
         learners: state.learners,
         loading: action.loading,
+        skills: state.skills,
         skillSearch: state.skillSearch,
         nameSearch: state.nameSearch,
       };
